backend/pyproject.toml
[project]
name = "pump_bot"
version = "2.0"
description = "Trade tokens on pump.fun"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "base58>=2.1.1",
    "borsh-construct>=0.1.0",
    "construct>=2.10.67",
    "construct-typing>=0.5.2",
    "solana==0.36.6",
    "solders>=0.26.0",
    "websockets>=15.0",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.11.13",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "protobuf>=5.29.4",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.10.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0"
]
performance = [
    "uvloop>=0.21.0",  # Optional: Better async performance (Linux/macOS only, not Windows)
]
web = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.9.0",
    "pynacl>=1.5.0",  # For signature verification
    "verify-crypto-signature>=0.0.1",  # Solana-specific signature verification
]
ai = [
    "openai>=1.12.0",
    "anthropic>=0.18.0",
]
social = [
    "tweepy>=4.14.0",  # Twitter API
    "telethon>=1.34.0",  # Telegram API
    "discord.py>=2.3.0",  # Discord API
    "praw>=7.7.0",  # Reddit API
    "beautifulsoup4>=4.12.0",  # Web scraping fallback
    "selenium>=4.15.0",  # Web scraping (optional)
]

[project.scripts]
pump_bot = "bot_runner:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = [
    "E", "F", "I", "UP", "N", "B", "A", "C4", "T10", "ARG", "PTH",
    "ANN",  # type annotations
    "S",    # security best practices
    "BLE",  # blind except statements
    "FBT",  # boolean trap parameters
    "C90",  # complexity metrics
    "TRY",  # exception handling best practices
    "SLF",  # private member access
    "TCH",  # type checking issues
    "RUF",  # Ruff-specific rules
    "ERA",  # eradicate commented-out code
    "PL",   # pylint conventions
]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
